+++ astra-sim/workload/CSVWriter.hh
@@ -1,0 +1,55 @@
+/******************************************************************************
+Copyright (c) 2020 Georgia Institute of Technology
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
+*******************************************************************************/
+#ifndef __CSVWRITER_HH__
+#define __CSVWRITER_HH__
+
+
+#include <map>
+#include <math.h>
+#include <fstream>
+#include <chrono>
+#include <ctime>
+#include <tuple>
+#include <cstdint>
+#include <iostream>
+#include <sys/stat.h>
+#include <unistd.h>
+#include <fcntl.h>
+
+namespace AstraSim{
+class CSVWriter{
+public:
+    //std::fstream inFile;
+    std::fstream myFile;
+    void initialize_csv(int rows, int cols);
+    CSVWriter(std::string path,std::string name);
+    void write_cell(int row,int column,std::string data);
+    std::string path;
+    std::string name;
+    ~CSVWriter(){
+        myFile.close();
+    }
+    inline bool exists_test(const std::string& name) {
+        struct stat buffer;
+        return (stat (name.c_str(), &buffer) == 0);
+    }
+};
+#endif
